buddy/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
google-genai
python-dotenv
httpx
pydantic
